backend/requirements.txt
# FastAPI & server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis>=5.0.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation & env
pydantic==2.6.1
pydantic-settings==2.1.0

# Excel & AI
openpyxl==3.1.2
pandas==2.2.0
openai==1.12.0
httpx==0.26.0
rapidfuzz>=3.0.0

# File upload
python-multipart==0.0.9

# Tests
pytest==8.0.0
pytest-asyncio==0.23.4
